api_server/requirements.txt
fastapi[all]
uvicorn[standard]
boto3
python-dotenv
pydantic
pydantic[dotenv]
requests
ffmpeg-python
pymongo